pyproject.toml
[project]
name = "romulus"
version = "0.1.0"
description = "An agent operating system where agents are born, sleep, heal, evolve, and die."
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "chromadb>=0.5",
    "apscheduler>=3.10,<4.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pyyaml>=6.0",
    "keyring>=25.0",
    "python-telegram-bot>=21",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov",
    "ruff",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["romulus*"]

[tool.ruff]
target-version = "py312"
line-length = 120
